Even more painfully slow, but better compression. Huffman draft.

+++ SeikenDensetsu3/bytepair.hpp
@@ -1,10 +1,11 @@
 #pragma once
+#include <list>
 #include <vector>
 
 namespace BytePair
 {
 
-std::vector<uint32_t>* compress(std::vector<std::vector<uint8_t>*>& blocksData);
+std::vector<uint32_t>* compress(std::vector<std::vector<uint16_t>*>& blocksData);
 std::vector<uint8_t> decompress(uint16_t v, const uint16_t* meta);
 
 }
